[Update] eventEdit page and add updateImage method, and update deleteImage method

+++ src/hooks/useDeleteItems/index.jsx
@@ -1,4 +1,6 @@
 import { useState } from 'react';
+import { storage } from '../../firebase';
+import defaultImage from '../../assets/default-image.png';
 
 function useDeleteItems() {
   const [modal, setModal] = useState(false);
@@ -11,13 +13,29 @@ function useDeleteItems() {
     setModal(false);
   };
 
-  const deleteImage = (eventImage, storage) => {
-    let imageUrl = eventImage;
+  const deleteImage = (eventImage) => {
+    if (eventImage !== defaultImage) {
+      let imageUrl = eventImage;
 
-    storage
-      .refFromURL(imageUrl)
-      .delete()
-      .catch((err) => console.error(err));
+      storage
+        .refFromURL(imageUrl)
+        .delete()
+        .catch((err) => console.error(err));
+    }
+  };
+
+  const updateImage = (imageUrl, eventData) => {
+    let currentImage = eventData.eventImage;
+    let newImage = imageUrl;
+
+    if (currentImage) {
+      if (newImage !== '') {
+        storage
+          .refFromURL(currentImage)
+          .delete()
+          .catch((err) => console.log(err));
+      }
+    }
   };
 
   return {
@@ -26,6 +44,7 @@ function useDeleteItems() {
     handleOpenModal,
     handleCloseModal,
     deleteImage,
+    updateImage,
   };
 }
